Substantial changes to winio.c that appear to fix the display problems that we have been seeing on lines with tabs.
It introduces the concept of "pages" which might not be the best choice
for them.... Basically, it's the part of a line that would be visible on
screen at once.  the first part (visible when you open a file) is page
1... the second (you have to scroll to the right approx COLS, is 2.  The
third is at approx (2 * COLS) - 7 and so on...)

It's not terribly efficient as of yet, there are probabally quite a few
redundant calls to xpt(), for instance.


git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@6 35c25a1d-7b9e-4130-9fde-d3aeb78583b8